Modularized handlers and helpers in the server. Moved index.js to the Server folder. Corrected in inconsistency in the names of properties in the dummyData.js
