+++ Server/Databases/Models/dummyData.js
@@ -13,24 +13,24 @@ exports.methodsData = [
       descriptions: [
         {
           text:'efa',
-          votes: 10,
+          upvotes: 10,
           entryID: 123,
           additions: [
             {
               text: 'asdfsds',
-              votes: 3,
+              upvotes: 3,
               additionID: 12
             }
           ]
         },
          {
           text:'asdfds',
-          votes: 12,
+          upvotes: 12,
           entryID: 124,
           additions: [
             {
               text: 'asdf',
-              votes: 3,
+              upvotes: 3,
               additionID: 11
             }
           ]
@@ -39,12 +39,12 @@ exports.methodsData = [
       examples: [
         {
           text:'asdfds',
-          votes: 12,
+          upvotes: 12,
           entryID: 125,
           additions: [
             {
               text: 'asdf',
-              votes: 3,
+              upvotes: 3,
               additionID: 10
             }
           ]
@@ -53,12 +53,12 @@ exports.methodsData = [
       tips: [
         {
           text:'asgsrdgdrgsdgdfds',
-          votes: 12,
+          upvotes: 12,
           entryID: 104,
           additions: [
             {
               text: 'rere',
-              votes: 3,
+              upvotes: 3,
               additionID: 129
             }
           ]
@@ -80,24 +80,24 @@ exports.methodsData = [
       descriptions: [
         {
           text:'efa',
-          votes: 14,
+          upvotes: 14,
           entryID: 1223,
           additions: [
             {
               text: 'asdf',
-              votes: 3,
+              upvotes: 3,
               additionID: 13
             }
           ]
         },
          {
           text:'bvb',
-          votes: 12,
+          upvotes: 12,
           entryID: 128,
           additions: [
             {
               text: 'asdf',
-              votes: 3,
+              upvotes: 3,
               additionID: 14
             }
           ]
@@ -106,12 +106,12 @@ exports.methodsData = [
       examples: [
         {
           text:'asdfds',
-          votes: 12,
+          upvotes: 12,
           entryID: 115,
           additions: [
             {
               text: 'asdf',
-              votes: 3,
+              upvotes: 3,
               additionID: 11
             }
           ]
@@ -120,12 +120,12 @@ exports.methodsData = [
       tips: [
         {
           text:'asgsrdgdrgsdgdfds',
-          votes: 12,
+          upvotes: 12,
           entryID: 154,
           additions: [
             {
               text: 'rere',
-              votes: 3,
+              upvotes: 3,
               additionID: 125
             }
           ]
