+++ Server/app.js
@@ -1,10 +1,9 @@
 var express = require('express');
 var path = require('path');
-var favicon = require('serve-favicon');
+var favicon = require('serve-favicon'); // currently not used
 var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
-var mongoose = require('mongoose');
 
 //var crowdsourceRouter = require('./Routes/crowdsource');
 var usersRouter = require('./Routes/users');
@@ -22,7 +21,7 @@ app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'StaticPages')));
+app.use(express.static(path.join(__dirname, 'StaticPages'))); // will this be used?
 
 app.use('/users', usersRouter); // might change later to not use router
